add requires layout tests, fix tests for linux
